requirements.txt
streamlit
langchain
langchain-community
langchain-chroma
langchain-ollama
langchain-groq
langchain-openai
langchain-huggingface
langchain-text-splitters
chromadb
python-dotenv
pypdf
huggingface-hub
sentence-transformers
pysqlite3-binary
fastapi
uvicorn
requests
